requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
python-dotenv>=1.0.1
python-slugify>=8.0.4
